pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "SmartShop"
license = { text = 'SmartBase' }
dependencies = [
    "django~=5.0.3",
    "django-post-office~=3.7.0",
    # temporary dependency, waiting for new release
    "django-dynamic-preferences@git+https://github.com/agateblue/django-dynamic-preferences.git@69f585b793e33de38ec678b5e56491c42ddf6f64",
    "django-smartbase-admin@git+https://github.com/SmartBase-SK/django-smartbase-admin.git@2299a5928b7fadc617abcdf1b720f70163c30ae9",
    "django-redis~=5.2.0",
    "redis~=4.5.5",
    "psutil~=5.9.5",
    "pyinstrument~=4.6.1",
    "requests~=2.31.0",
    "requests-oauthlib~=1.3.1",
    "django-health-check~=3.17.0",
    "psycopg[binary]~=3.1.9",
    "celery>=5.4.0",
    "django-celery-results~=2.5.1",
    "django-sekizai~=4.1.0",
    "python-dotenv>=1.0.1",
    "setuptools>=69.2.0",
]


readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true
[tool.setuptools]
py-modules = []
